# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Agent Data Gateway",
    "description": "Quant Agent 执行引擎 HTTP/WS/SSE 契约（BE-19）。\n\n## 统一响应\n\n```json\n{\"code\": 0, \"msg\": \"ok\", \"data\": {}, \"ts\": 1719475200000}\n```\n\n- `code=0` 成功；非零见错误码（`docs/10` §1.4）\n- 响应头 `X-Trace-Id` 用于全链路追踪（BE-10）\n\n## 机器可读契约\n\n导出产物：`docs/openapi.json`（`python scripts/export_openapi.py`）。\n人工互校文档：`docs/10. API接口规范.md`。",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/system/data-quality": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Data Quality",
        "description": "SVC-04 校验结果汇总：按数据源展示脏数据率 / 完整率 / 价格异常 / 过期计数。\nGrafana 独立面板订阅 Prometheus 同名指标；本接口供前端/运维即时查看。",
        "operationId": "get_data_quality_api_v1_system_data_quality_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/system/performance-logs": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Performance Logs",
        "description": "获取系统性能监控日志（慢请求与事件循环卡顿）",
        "operationId": "get_performance_logs_api_v1_system_performance_logs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "description": "返回条数上限",
              "default": 100,
              "title": "Limit"
            },
            "description": "返回条数上限"
          },
          {
            "name": "log_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按类型筛选: slow_request / event_loop_block",
              "title": "Log Type"
            },
            "description": "按类型筛选: slow_request / event_loop_block"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 时间戳，只返回此时间之后的日志",
              "title": "Since"
            },
            "description": "ISO 时间戳，只返回此时间之后的日志"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/performance-stats": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Performance Stats",
        "description": "返回指定时间窗口内的性能聚合统计",
        "operationId": "get_performance_stats_api_v1_system_performance_stats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "description": "统计时间窗口（小时）",
              "default": 24,
              "title": "Hours"
            },
            "description": "统计时间窗口（小时）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/apm-dashboard": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Apm Dashboard",
        "description": "一次请求返回 APM 面板所需的全部数据：\n- health: 组件健康状态\n- cluster: 集群拓扑\n- metrics: Prometheus 指标快照\n- performance_stats: 24h 性能统计",
        "operationId": "apm_dashboard_api_v1_system_apm_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}